% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{blanco_palettes}
\alias{blanco_palettes}
\alias{custom}
\alias{gradients}
\alias{grays}
\alias{misc_pal}
\title{Blanco palettes}
\format{
\subsection{\code{blanco_palettes}}{

An object of class \code{palettes_palette} with \verb{r length(blanco_palettes)}
color palettes. Use \code{names(blanco_palettes)} to return all palette names.
}

\subsection{\code{custom}}{

An object of class \code{palettes_palette} with
\verb{r length(custom)} sequential colour palettes.
Use \code{names(custom)} to return all palette names.
}

\subsection{\code{gradients}}{

An object of class \code{palettes_palette} with
\verb{r length(gradients)} discrete color palettes.
Use \code{names(gradients)} to return all palette names.
}

\subsection{\code{grays}}{

An object of class \code{palettes_palette} with
\verb{r length(grays)} discrete color palettes.
Use \code{names(grays)} to return all palette names.
}

\subsection{\code{misc_pal}}{

An object of class \code{palettes_palette} with
\verb{r length(misc_pal)} sequential colour palettes.
Use \code{names(misc_pal)} to return all palette names.
}
}
\source{
\url{https://github.com/claudia-laurenzano/blancocentR}
}
\usage{
blanco_palettes

custom

gradients

grays

misc_pal
}
\description{
Palettes for a cohesive look across Blanco Center visualizations.
}
\examples{
# Get all palettes by name.
names(blanco_palettes)

# Plot all palettes.
plot(blanco_palettes)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_color]{pal_color()}}
}
\author{
\href{https://github.com/claudia-laurenzano}{Claudia Laurenzano}
}
\keyword{datasets}
