% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blancocentR-package.R
\docType{package}
\name{blancocentR-package}
\alias{blancocentR}
\alias{blancocentR-package}
\title{blancocentR: A helper package for Blanco Center staff}
\description{
A ggplot2 theme for a cohesive look across Blanco Center visualizations. Color palettes to be used in Blanco Center visualizations. Palettes built using the 'palettes' package, which provides methods for printing, formatting, casting and coercion, extraction and updating of components, plotting, color mixing arithmetic, and color interpolation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/user/repo}
  \item Report bugs at \url{https://github.com/user/repo/issues}
}

}
\author{
\strong{Maintainer}: Claudia Laurenzano \email{claudialaurenzano@gmail.com}

}
\keyword{internal}
