% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{col_la}
\alias{col_la}
\title{Louisiana blue}
\format{
A character string containing the hex code for Louisiana blue: "#4776aa".
}
\usage{
col_la
}
\description{
Named color used for Louisiana-related data.
}
\keyword{datasets}
