% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{col_ru}
\alias{col_ru}
\title{Rural}
\format{
A character string containing the hex code for rural: "#23464f".
}
\usage{
col_ru
}
\description{
Named color used for rural data. Also included in the miscellaneous palette.
}
\keyword{datasets}
