% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{col_ur}
\alias{col_ur}
\title{Color: Urban}
\format{
A character string containing the hex code for urban: "#d24a41".
}
\usage{
col_ur
}
\description{
Named color used for urban data. Also included in the miscellaneous palette.
}
\keyword{datasets}
