% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{col_us}
\alias{col_us}
\title{U.S. red}
\format{
A character string containing the hex code for U.S. red: "#711F30".
}
\usage{
col_us
}
\description{
Named color used for U.S.-related data.
}
\keyword{datasets}
