% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{dcfs_pal}
\alias{dcfs_pal}
\title{Color palette for DCFS regions}
\usage{
dcfs_pal(region = "name", darken = 0)
}
\arguments{
\item{region}{System used to identify regions: 'number' or 'name'.}

\item{darken}{Factor by which the shade for region 9 - Monroe gets darkened. Default = 0.}
}
\value{
A character object of length 9.
}
\description{
Colors are defined by region number or region name. The shade for region
9 - Monroe can be darkened for small area geoms (e.g., geom_point()).
}
\examples{
#' \dontrun{
dcfs_pal()
dcfs_pal(region = "number")
dcfs_pal(region = "number", darken = 1)
}
}
