% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{de_palette}
\alias{de_palette}
\title{Color palette for Digital Equity regions (LAPDD)}
\usage{
de_palette(region = "name", darken = 0)
}
\arguments{
\item{region}{System used to identify regions: 'number' or 'name'.}

\item{darken}{Factor by which the shade for region 8 - 8 - North Delta gets darkened. Default = 0.}
}
\value{
A character object of length 8.
}
\description{
Colors are defined by region number or region name. The shade for region
8 - North Delta can be darkened for small area geoms (e.g., geom_point()).
Regions via the Louisiana Association of Planning and Development District.
}
\examples{
\dontrun{
de_palette()
de_palette(region = "number", darken = 1)
}
}
