% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louisiana_geography.R
\name{get_region_geometry}
\alias{get_region_geometry}
\title{Title}
\usage{
get_region_geometry(df, region_number, region_name)
}
\arguments{
\item{df}{A dataframe or tibble.}

\item{region_number}{Column with numeric or alphanumeric region ID}

\item{region_name}{Column with region name ID}
}
\value{
A dataframe or tibble.
}
\description{
Title
}
\examples{
get_region_geometry(louisiana, region_DCFS, region_name_DCFS)
}
