% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{gradients}
\alias{gradients}
\title{Color gradient palettes function for general purpose use, using colorspace::lighten}
\usage{
gradients(colors, n = 11)
}
\arguments{
\item{colors}{List of colors to be lightened.}

\item{n}{Number of shades in the palette. Default = 11 (lighten by 10\% steps).}
}
\value{
A list.
}
\description{
Create gradient color palettes.
}
\examples{
\dontrun{
gradients(race_pal)
}
}
