% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_bc.R
\name{gt_theme_bc}
\alias{gt_theme_bc}
\title{A gt theme for Blanco Center tables}
\usage{
gt_theme_bc(
  gt,
  base_family = "Quicksand",
  header_family = "Bitter",
  text_color = "#575E66",
  title_color = "#a00b0b",
  base_size = 12,
  ...
)
}
\arguments{
\item{gt}{An existing gt table object of class \code{gt_tbl}}

\item{base_family}{Font used for most text.}

\item{header_family}{Font used for titles: table title, column headers, spanner column headers, row group labels.}

\item{text_color}{Color used for everything else, except table title.}

\item{title_color}{Color used for table title.}

\item{base_size}{Base text size in pts (converted to px within function).}

\item{...}{Additional arguments passed to gt::tab_options().}
}
\value{
A gt theme.
}
\description{
\code{gt_theme_bc()} adds the Blanco Center theme to any table created using \code{gt()}.
Customize table elements like font family, or text color and sizes.
}
\examples{
\dontrun{
gt_theme_bc(header_family = "Bitter", base_family = "Quicksand")
}
}
