% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{la_gradient_pal}
\alias{la_gradient_pal}
\title{Color gradient palette for Louisiana}
\usage{
la_gradient_pal(base = "#1C2F43", n = 7)
}
\arguments{
\item{base}{Base color to be blended with "white". Default = "#1C2F43".}

\item{n}{Number of shades in the palette. Default = 7.}
}
\value{
A character object.
}
\description{
Create a gradient color for Louisiana.
}
\examples{
\dontrun{
la_gradient_pal()
la_gradient(n = 5)
}
}
