% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{lafirst_blue2}
\alias{lafirst_blue2}
\title{Color: LAFIRST blue}
\format{
A character string containing the hex code for LA FIRST blue: "#475ab3".
}
\usage{
lafirst_blue2
}
\description{
Named color used for titles in theme_lafirst() unless otherwise specified.
}
\keyword{datasets}
