% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{lafirst_pal}
\alias{lafirst_pal}
\title{LA FIRST palette}
\format{
A character vector with named color values.
}
\usage{
lafirst_pal
}
\description{
A named vector of hex color codes that define two additional colors for use in LA FIRST visualizations (additional to misc_pal).
}
\keyword{datasets}
