% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ldh_pal}
\alias{ldh_pal}
\title{Color palette for LDH regions}
\usage{
ldh_pal(region = "name", darken = 0)
}
\arguments{
\item{region}{System used to identify regions: 'number' or 'name'.}

\item{darken}{Factor by which the shade for region 8 - Northeast Louisiana gets darkened. Default = 0.}
}
\value{
A character object of length 9.
}
\description{
Colors are defined by region number or region name. The shade for region
8 - Northeast Louisiana can be darkened for small area geoms (e.g., geom_point()).
}
\examples{
#' \dontrun{
ldh_pal()
ldh_pal(region = "number")
ldh_pal(region = "number", darken = 1)
}
}
