% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{light_gray}
\alias{light_gray}
\title{Light gray}
\format{
A character string containing the hex code for Louisiana blue: "#D1D3D5".
}
\usage{
light_gray
}
\description{
Named color used for de-emphasized data and background plot elements.
}
\keyword{datasets}
