% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{light_text}
\alias{light_text}
\title{Light text}
\format{
A character string containing the hex code for Louisiana blue: "#575E66".
}
\usage{
light_text
}
\description{
Named color used for light text, used for axis lines, ticks, and all text in theme_bc() unless otherwise specified.
}
\keyword{datasets}
