% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{louisiana_long}
\alias{louisiana_long}
\title{Louisiana's 64 parishes, FIPS codes, and region designations according
to different state agencies and organizations, parish geometry, and parish centers.}
\format{
A data frame/sf object with 393 rows and 16 columns:
\describe{
\item{GEOID, FIPS}{Parish FIPS code in character & numeric formats}
\item{NAME}{Full parish name including "Parish, Louisiana"}
\item{variable}{Variable name for population density, consists of Census table and row number}
\item{estimate}{Estimate}
\item{moe}{Margin of error}
\item{year}{Year for which ACS-5 Year estimates were retrieved}
\item{parish, parish_full}{Parish name with & without "Parish"}
\item{metro_BLS}{"Metropolitan area" or "Nonmetropolitan area" as designated by BLS}
\item{pop_density}{Rural/Urban, threshold = 100000}
\item{org}{Name of agency or organization for region designation}
\item{region_number}{Region numeric ID}
\item{region_name}{Region name ID}
\item{geometry}{MULTIPOLYGON, parish outline}
\item{center}{POINT, parish center point}
}
}
\source{
U.S. Census Bureau, 'tidycensus' package, see sources for 'parishes'
}
\usage{
louisiana_long
}
\description{
Louisiana's 64 parishes, FIPS codes, and region designations according
to different state agencies and organizations, parish geometry, and parish centers.
}
\keyword{datasets}
