% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{mid_gray}
\alias{mid_gray}
\title{Color: Mid gray}
\format{
A character string containing the hex code for mid gray: "#94989D".
}
\usage{
mid_gray
}
\description{
Named color used for de-emphasized data.
}
\keyword{datasets}
