% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{misc_pal}
\alias{misc_pal}
\title{Miscellaneous Palette}
\format{
A character vector with named color values.
}
\usage{
misc_pal
}
\description{
A named vector of hex color codes that define miscellaneous colors for use in Blanco Center visualizations.
}
\keyword{datasets}
