% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parishes_long}
\alias{parishes_long}
\title{Louisiana's 64 parishes, FIPS codes, and region designations according
to different state agencies and organizations (long format).}
\format{
A data frame with 393 rows and 8 columns:
\describe{
\item{FIPS}{Parish FIPS code}
\item{parish, parish_full}{Parish name with & without "Parish"}
\item{name}{Full parish name including "Parish, Louisiana"}
\item{metro_BLS}{"Metropolitan area" or "Nonmetropolitan area" as designated by BLS}
\item{org}{Name of agency or organization for region designation}
\item{region_number}{Region numeric ID}
\item{region_name}{Region name ID}
}
}
\source{
LAPDD, DCFS, One Acadiana, LWC, BLS, LDH, OJJ
}
\usage{
parishes_long
}
\description{
Louisiana's 64 parishes, FIPS codes, and region designations according
to different state agencies and organizations (long format).
}
\keyword{datasets}
