% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{race_muted_pal}
\alias{race_muted_pal}
\title{LA FIRST muted race/ethnicity palette}
\format{
A character vector with named color values.
}
\usage{
race_muted_pal
}
\description{
A named vector of hex color codes for use in LA FIRST visualizations showing race/ethnicity by color. A more muted selection
}
\keyword{datasets}
