% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ru_gradient_pal}
\alias{ru_gradient_pal}
\title{Color gradient palette for rural areas.}
\usage{
ru_gradient_pal(base = "#23464f", n = 7)
}
\arguments{
\item{base}{Base color to go darker and lighter. Default = "#23464f".}

\item{n}{Number of shades in the palette. Default = 7.}
}
\value{
A character object.
}
\description{
Create a gradient color for rural areas.
}
\examples{
#' \dontrun{
ru_gradient_pal()
}
}
