% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{sex_gender_pal}
\alias{sex_gender_pal}
\title{LA FIRST sex/gender palette}
\format{
A character vector with named color values.
}
\usage{
sex_gender_pal
}
\description{
A named vector of hex color codes for use in LA FIRST visualizations showing sex or gender by color.
}
\keyword{datasets}
