% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{show_cvd}
\alias{show_cvd}
\title{Simulate color vision deficiency}
\usage{
show_cvd(x)
}
\arguments{
\item{x}{A color, a vector of colors, or a color palette.}
}
\value{
A plot showing selected color or colors in different color deficiency simulations.
}
\description{
Simulate color(s) to check how people with different variations of
colorblindness may perceive your visualization. Supported variations include
deuteranopia, protanopia, tritanopia, and grayscale.
}
\examples{
x <- c("steelblue", "firebrick", "gold")
show_cvd(x)
}
