% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bc.R
\name{theme_bc}
\alias{theme_bc}
\title{Title}
\usage{
theme_bc(
  title_font = "Bitter",
  base_font = "Quicksand",
  dark_text = "#1A242F",
  light_text = "#575E66",
  void = FALSE,
  grid_x = FALSE,
  grid_y = FALSE,
  base_size = 12,
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{title_font}{Font used for titles: plot title, axis titles, legend title.}

\item{base_font}{Font used for all other text.}

\item{dark_text}{Color used for axis and legend titles.}

\item{light_text}{Color used for everything else, except plot title.}

\item{void}{TRUE or FALSE: removes all axis elements. Suggested for maps.}

\item{grid_x}{TRUE or FALSE: adds x gridlines.}

\item{grid_y}{TRUE or FALSE: adds y gridlines.}

\item{base_size}{Base text size.}

\item{base_line_size}{Base line size.}

\item{base_rect_size}{Base rectangle size.}
}
\value{
A ggplot2 theme.
}
\description{
Title
}
\examples{
theme_bc(title_font = "Bitter", base_font = "Quicksand", void = TRUE)
}
