% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bc_legacy.R
\name{theme_bc_legacy}
\alias{theme_bc_legacy}
\title{A ggplot2 theme for Blanco Center visualizations}
\usage{
theme_bc_legacy(
  title_font = "Bitter",
  base_font = "Quicksand",
  dark_text = "#1A242F",
  light_text = "#575E66",
  void = FALSE,
  grid_x = FALSE,
  grid_y = FALSE,
  base_size = 12,
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{title_font}{Font used for titles: plot title, axis titles, legend title.}

\item{base_font}{Font used for all other text.}

\item{dark_text}{Color used for axis and legend titles.}

\item{light_text}{Color used for everything else, except plot title.}

\item{void}{TRUE or FALSE: removes all axis elements. Suggested for maps.}

\item{grid_x}{TRUE or FALSE: adds x gridlines.}

\item{grid_y}{TRUE or FALSE: adds y gridlines.}

\item{base_size}{Base text size in pts.}

\item{base_line_size}{Base line size in pts.}

\item{base_rect_size}{Base rectangle size in pts.}
}
\value{
A ggplot2 theme.
}
\description{
\code{theme_bc_legacy()} adds the Blanco Center theme to any plot created using \code{ggplot()}.
Customize plot elements like font family and color and axis colors and line sizes.
Grid lines for x and y axes can be added, or all axes can be removed, e.g. for maps.
Note: This is the legacy version of the Blanco Center theme function (developed in 2023).
For the updated version, use theme_bc(), developed Fall 2025.
}
\examples{
\dontrun{
theme_bc_legacy(title_font = "Bitter", base_font = "Quicksand", void = TRUE)
}
}
