% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lafirst.R
\name{theme_lafirst}
\alias{theme_lafirst}
\title{A ggplot2 theme for Blanco Center LA FIRST visualizations}
\usage{
theme_lafirst(
  title_font = "Univers LT Std",
  base_font = "Univers LT Std",
  dark_text = "#1A242F",
  light_text = "#575E66",
  void = FALSE,
  grid_x = FALSE,
  grid_y = FALSE,
  base_size = 8,
  base_line_size = base_size/40,
  base_rect_size = base_size/40
)
}
\arguments{
\item{title_font}{Font used for titles: plot title, axis titles, legend title.}

\item{base_font}{Font used for all other text.}

\item{dark_text}{Color used for axis and legend titles.}

\item{light_text}{Color used for everything else, except plot title.}

\item{void}{TRUE or FALSE: if TRUE, removes all axis elements. Suggested for maps.}

\item{grid_x}{TRUE or FALSE: if TRUE, adds x grid lines.}

\item{grid_y}{TRUE or FALSE: if TRUE, adds y grid lines.}

\item{base_size}{Base text size in pts.}

\item{base_line_size}{Base line size in pts.}

\item{base_rect_size}{Base rectangle size in pts.}
}
\value{
A ggplot2 theme.
}
\description{
\code{theme_lafirst()} adds the LA FIRST theme to any plot created using \code{ggplot()}.
Customize plot elements like font family and color and axis colors and line sizes.
Grid lines for x and y axes can be added, or all axes can be removed, e.g. for maps.
}
\examples{
\dontrun{
theme_lafirst(title_font = "Univers LT Std", base_font = "Univers LT Std", void = TRUE)
}
}
