% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{ull_red}
\alias{ull_red}
\title{Color: UL red}
\format{
A character string containing the hex code for UL red: "#a00b0b".
}
\source{
\href{https://louisiana.edu/web/tutorials-resources/web-style-guide/approved-web-colors}{UL approved web colors}
}
\usage{
ull_red
}
\description{
Named color used for UL Vermillion red, used as title color in theme_bc().
}
\keyword{datasets}
