% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ur_gradient_palette}
\alias{ur_gradient_palette}
\title{Color gradient palette for urban areas.}
\usage{
ur_gradient_palette(base = "#d24a41", n = 7)
}
\arguments{
\item{base}{Base color to go darker and lighter. Default = "#d24a41".}

\item{n}{Number of shades in the palette. Default = 7.}
}
\value{
A character object.
}
\description{
Create a gradient color for urban areas.
}
\examples{
\dontrun{
ur_gradient_palette()
ur_gradient_palette(n = 5)
}
}
