% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{us_gradient_pal}
\alias{us_gradient_pal}
\title{Color gradient palette for the U.S.}
\usage{
us_gradient_pal(base = "#5A1826", n = 7)
}
\arguments{
\item{base}{Base color to be blended with "white". Default = "#5A1826".}

\item{n}{Number of shades in the palette. Default = 7.}
}
\value{
A character object.
}
\description{
Create a gradient color for the U.S.
}
\examples{
#' \dontrun{
us_gradient_pal()
}
}
