% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{blanco_palettes}
\alias{blanco_palettes}
\alias{custom}
\alias{gradients}
\alias{grays}
\alias{misc_pal}
\alias{cols_dcfs}
\alias{cols_dcfs1}
\alias{cols_dcfs1_nm}
\alias{cols_dcfs2}
\alias{cols_dcfs2_nm}
\alias{cols_dcfs_nm}
\alias{cols_de}
\alias{cols_de1}
\alias{cols_de1_nm}
\alias{cols_de2}
\alias{cols_de2_nm}
\alias{cols_de_nm}
\alias{cols_ldh}
\alias{cols_ldh1}
\alias{cols_ldh1_nm}
\alias{cols_ldh2}
\alias{cols_ldh2_nm}
\alias{cols_ldh_nm}
\title{Blanco palettes}
\format{
\subsection{\code{blanco_palettes}}{

An object of class \code{palettes_palette} with 190
color palettes. Use \code{names(blanco_palettes)} to return all palette names.
}

\subsection{\code{custom}}{

An object of class \code{palettes_palette} with
5 discrete colors.
Use \code{names(custom)} to return all color names.
}

\subsection{\code{gradients}}{

An object of class \code{palettes_palette} with
4 continuous color palettes.
Use \code{names(gradients)} to return all palette names.
}

\subsection{\code{grays}}{

An object of class \code{palettes_palette} with
4 discrete colors.
Use \code{names(grays)} to return all color names.
}

\subsection{\code{misc_pal}}{

An object of class \code{palettes_palette} with
21 discrete colors.
Use \code{names(misc_pal)} to return all color names.
}

\subsection{\code{cols_dcfs}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_dcfs)} to return all color names.
}

\subsection{\code{cols_dcfs1}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_dcfs1)} to return all color names.
}

\subsection{\code{cols_dcfs1_nm}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_dcfs1_nm)} to return all color names.
}

\subsection{\code{cols_dcfs2}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_dcfs2)} to return all color names.
}

\subsection{\code{cols_dcfs2_nm}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_dcfs2_nm)} to return all color names.
}

\subsection{\code{cols_dcfs_nm}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_dcfs_nm)} to return all color names.
}

\subsection{\code{cols_de}}{

An object of class \code{palettes_palette} with
8 discrete colors.
Use \code{names(cols_de)} to return all color names.
}

\subsection{\code{cols_de1}}{

An object of class \code{palettes_palette} with
8 discrete colors.
Use \code{names(cols_de1)} to return all color names.
}

\subsection{\code{cols_de1_nm}}{

An object of class \code{palettes_palette} with
8 discrete colors.
Use \code{names(cols_de1_nm)} to return all color names.
}

\subsection{\code{cols_de2}}{

An object of class \code{palettes_palette} with
8 discrete colors.
Use \code{names(cols_de2)} to return all color names.
}

\subsection{\code{cols_de2_nm}}{

An object of class \code{palettes_palette} with
8 discrete colors.
Use \code{names(cols_de2_nm)} to return all color names.
}

\subsection{\code{cols_de_nm}}{

An object of class \code{palettes_palette} with
8 discrete colors.
Use \code{names(cols_de_nm)} to return all color names.
}

\subsection{\code{cols_ldh}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_ldh)} to return all color names.
}

\subsection{\code{cols_ldh1}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_ldh1)} to return all color names.
}

\subsection{\code{cols_ldh1_nm}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_ldh1_nm)} to return all color names.
}

\subsection{\code{cols_ldh2}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_ldh2)} to return all color names.
}

\subsection{\code{cols_ldh2_nm}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_ldh2_nm)} to return all color names.
}

\subsection{\code{cols_ldh_nm}}{

An object of class \code{palettes_palette} with
9 discrete colors.
Use \code{names(cols_ldh_nm)} to return all color names.
}
}
\source{
\url{https://github.com/claudia-laurenzano/blancocentR}
}
\usage{
blanco_palettes

custom

gradients

grays

misc_pal

cols_dcfs

cols_dcfs1

cols_dcfs1_nm

cols_dcfs2

cols_dcfs2_nm

cols_dcfs_nm

cols_de

cols_de1

cols_de1_nm

cols_de2

cols_de2_nm

cols_de_nm

cols_ldh

cols_ldh1

cols_ldh1_nm

cols_ldh2

cols_ldh2_nm

cols_ldh_nm
}
\description{
Palettes for a cohesive look across Blanco Center visualizations.
}
\examples{
# Get all palettes by name.
names(blanco_palettes)

# Plot all palettes.
plot(blanco_palettes)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_color]{pal_color()}}
}
\author{
\href{https://github.com/claudia-laurenzano}{Claudia Laurenzano}
}
\keyword{datasets}
