% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blancocentR-package.R
\docType{package}
\name{blancocentR-package}
\alias{blancocentR}
\alias{blancocentR-package}
\title{blancocentR: A helper package for Blanco Center staff}
\description{
A ggplot2 theme for a cohesive look across Blanco Center visualizations. Color palettes to be used in Blanco Center visualizations. Palettes built using the 'palettes' package, which provides methods for printing, formatting, casting and coercion, extraction and updating of components, plotting, color mixing arithmetic, and color interpolation.Louisiana geography (parishes and regions) for easy mapping. Shapefiles and population data from the U.S. Census aquired via the 'tidycensus' package.Includes 'parishes' and 'louisiana' data: Louisiana's 64 parishes, region designations via different state agencies and organizations, and geometry for parishes and regions, as well as center points.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/claudia-laurenzano/blancocentR/tree/main}
  \item Report bugs at \url{https://github.com/claudia-laurenzano/blancocentR/issues}
}

}
\author{
\strong{Maintainer}: Claudia Laurenzano \email{claudialaurenzano[at]gmail.com}

}
\keyword{internal}
