% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{dark_text}
\alias{dark_text}
\title{Dark text}
\format{
A character string containing the hex code for dark text: "#1A242F".
}
\usage{
dark_text
}
\description{
Named color used for dark text, used for axis, legend, and facet titles in theme_bc().
}
\keyword{datasets}
