% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louisiana_geography.R
\name{get_region_geometry}
\alias{get_region_geometry}
\title{Summarize parish-level geometry by region}
\usage{
get_region_geometry(df, region_number, region_name)
}
\arguments{
\item{df}{An sf object}

\item{region_number}{Column with numeric or alphanumeric region ID}

\item{region_name}{Column with region name ID}
}
\value{
An sf object
}
\description{
\code{get_region_geometry()} summarizes parish polygons to a region polygon and
returns an sf object with the region identifier(s) based on your region
input, a geometry column containing the region polygon, and a center column
with the region's center.
}
\examples{
\dontrun{
get_region_geometry(louisiana, region_DCFS, region_name_DCFS)
}
}
