% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{misc_pal}
\alias{misc_pal}
\title{Color palette for miscellaneous uses}
\format{
An object of class \code{list} of length 21.
}
\usage{
misc_pal
}
\value{
A character of length 21, a list of length 21.
}
\description{
Create a color palette with 21 colors. Please note that misc_pal$vintageviolet
("#77839b") is used in some regional color palettes (e.g., DCFS, LDH).
}
\examples{
\dontrun{
misc_pal()
misc_pal(get_list = TRUE)
}
}
\keyword{datasets}
