% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bc.R
\name{show_cvd}
\alias{show_cvd}
\title{Title}
\usage{
show_cvd(x)
}
\arguments{
\item{x}{A color, a vector of colors, or a color palette.}
}
\value{
A plot showing selected color or colors in different color deficiency simulations.
}
\description{
Title
}
\examples{
x <- c("steelblue", "firebrick", "gold")
show_cvd(x)
}
